// src/server/api/constants.ts
export enum SubjectNodeType {
  CHAPTER = "CHAPTER",
  TOPIC = "TOPIC",
  SUBTOPIC = "SUBTOPIC"
}

export enum CompetencyLevel {
  BASIC = "BASIC",
  INTERMEDIATE = "INTERMEDIATE",
  ADVANCED = "ADVANCED",
  EXPERT = "EXPERT"
}